% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregation.R
\name{.pairwise_bhattacharyya}
\alias{.pairwise_bhattacharyya}
\title{Pairwise Bhattacharyya coefficients between Beta distributions
Extended numerically for other families.}
\usage{
.pairwise_bhattacharyya(priors)
}
\description{
Pairwise Bhattacharyya coefficients between Beta distributions
Extended numerically for other families.
}
