% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot.bayprior}
\alias{plot.bayprior}
\title{Plot method for bayprior objects}
\usage{
\method{plot}{bayprior}(x, ...)
}
\arguments{
\item{x}{A \code{bayprior} object from any \code{elicit_*()} function.}

\item{...}{Additional arguments (currently unused).}
}
\value{
A \code{ggplot} object.
}
\description{
Produces a density plot of a fitted prior distribution with the 95\%
credible interval shaded and the prior mean shown as a dashed vertical
line. Dispatched automatically when \code{plot()} is called on a
\code{bayprior} object.
}
\examples{
prior <- elicit_beta(mean = 0.30, sd = 0.10, method = "moments",
                     label = "Response rate")
plot(prior)

}
