% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot_sensitivity}
\alias{plot_sensitivity}
\title{Plot sensitivity analysis results}
\usage{
plot_sensitivity(sensitivity, target = NULL, highlight_reference = TRUE)
}
\arguments{
\item{sensitivity}{A \code{bayprior_sensitivity} object.}

\item{target}{Character. Which target quantity to plot.}

\item{highlight_reference}{Logical. Default \code{TRUE}.}
}
\value{
A \code{ggplot} object.
}
\description{
Plot sensitivity analysis results
}
