% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot_tornado}
\alias{plot_tornado}
\title{Tornado plot of prior influence on posterior quantities}
\usage{
plot_tornado(sensitivity, title = "Prior influence on posterior estimates")
}
\arguments{
\item{sensitivity}{A \code{bayprior_sensitivity} object.}

\item{title}{Character. Plot title.}
}
\value{
A \code{ggplot} object.
}
\description{
Tornado plot of prior influence on posterior quantities
}
