% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_app.R
\name{run_app}
\alias{run_app}
\title{Run the bayprior Shiny Application}
\usage{
run_app(
  onStart = NULL,
  options = list(),
  enableBookmarking = NULL,
  uiPattern = "/",
  ...
)
}
\arguments{
\item{onStart}{A function called before the app runs.}

\item{options}{List of options passed to \code{shiny::shinyApp()}.}

\item{enableBookmarking}{Passed to \code{shiny::shinyApp()}.}

\item{uiPattern}{Passed to \code{shiny::shinyApp()}.}

\item{...}{Additional arguments passed to \code{golem::with_golem_options()}.}
}
\value{
A \code{shiny.appobj}, invisibly.
}
\description{
Launches the golem-structured shinydashboard application.
}
