% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{summary.bayprior}
\alias{summary.bayprior}
\title{Summary method for bayprior objects}
\usage{
\method{summary}{bayprior}(object, ...)
}
\arguments{
\item{object}{A \code{bayprior} object.}

\item{...}{Ignored.}
}
\value{
A list with summary statistics (invisibly).
}
\description{
Summary method for bayprior objects
}
