% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/robust_priors.R
\name{plot.bayprior_power_prior}
\alias{plot.bayprior_power_prior}
\title{Plot calibration curve for power prior weight selection}
\usage{
\method{plot}{bayprior_power_prior}(x, ...)
}
\arguments{
\item{x}{A \code{bayprior_power_prior} object.}

\item{...}{Ignored.}
}
\value{
A \code{ggplot} object.
}
\description{
Plot calibration curve for power prior weight selection
}
