% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{print.bayprior}
\alias{print.bayprior}
\title{Print method for bayprior objects}
\usage{
\method{print}{bayprior}(x, ...)
}
\arguments{
\item{x}{A \code{bayprior} object.}

\item{...}{Ignored.}
}
\description{
Print method for bayprior objects
}
