% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{print.bayprior_conflict}
\alias{print.bayprior_conflict}
\title{Print method for bayprior_conflict objects}
\usage{
\method{print}{bayprior_conflict}(x, ...)
}
\arguments{
\item{x}{A \code{bayprior_conflict} object.}

\item{...}{Ignored.}
}
\description{
Print method for bayprior_conflict objects
}
