% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conflict_mahalanobis.R
\name{print.bayprior_conflict_mv}
\alias{print.bayprior_conflict_mv}
\title{Print method for multivariate conflict objects}
\usage{
\method{print}{bayprior_conflict_mv}(x, ...)
}
\arguments{
\item{x}{A \code{bayprior_conflict_mv} object.}

\item{...}{Ignored.}
}
\description{
Print method for multivariate conflict objects
}
