% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/robust_priors.R
\name{print.bayprior_power_prior}
\alias{print.bayprior_power_prior}
\title{Print method for bayprior_power_prior objects}
\usage{
\method{print}{bayprior_power_prior}(x, ...)
}
\arguments{
\item{x}{A \code{bayprior_power_prior} object.}

\item{...}{Ignored.}
}
\description{
Print method for bayprior_power_prior objects
}
